Move `ByteCost` tests
