+++ bee-message/src/output/byte_cost.rs
@@ -103,50 +103,3 @@ impl<T: ByteCost, const N: usize> ByteCost for [T; N] {
         self.iter().map(|elem| elem.weighted_bytes(config)).sum()
     }
 }
-
-#[cfg(test)]
-mod test {
-    use crate::output::{ByteCost, ByteCostConfig, ByteCostConfigBuilder, Output};
-
-    use bee_test::rand::output::{rand_alias_output, rand_basic_output, rand_foundry_output, rand_nft_output};
-
-    use packable::{Packable, PackableExt};
-
-    const BYTE_COST: u64 = 1;
-    const FACTOR_KEY: u64 = 10;
-    const FACTOR_DATA: u64 = 1;
-
-    fn config() -> ByteCostConfig {
-        ByteCostConfigBuilder::new()
-            .byte_cost(BYTE_COST)
-            .key_factor(FACTOR_KEY)
-            .data_factor(FACTOR_DATA)
-            .finish()
-    }
-
-    // We have to jump through hoops here because the randomly generated outputs from `bee_test` have a different
-    // type then the Outputs of this crate (although they are technically the same).
-    fn convert<T: PackableExt>(rand_output: impl Packable) -> T {
-        let bytes = rand_output.pack_to_vec();
-        // Safety: We know it's the right type.
-        T::unpack_unverified(bytes).unwrap()
-    }
-
-    fn output_in_range(output: Output, range: std::ops::RangeInclusive<u64>) {
-        let cost = output.byte_cost(&config());
-        assert!(
-            range.contains(&cost),
-            "{:#?} has a required byte cost of {}",
-            output,
-            cost
-        );
-    }
-
-    #[test]
-    fn valid_byte_cost_range() {
-        output_in_range(Output::Alias(convert(rand_alias_output())), 445..=29_620);
-        output_in_range(Output::Basic(convert(rand_basic_output())), 414..=13_485);
-        output_in_range(Output::Foundry(convert(rand_foundry_output())), 496..=21_365);
-        output_in_range(Output::Nft(convert(rand_nft_output())), 435..=21_734);
-    }
-}
